server/requirements.txt
fastapi
uvicorn
neo4j
chromadb
python-multipart
python-dotenv
pydantic
pypdf
langchain
langchain-community
sentence-transformers
requests
